requirements.txt
diskcache
torch
transformers
gunicorn
flask
simpletransformers
pandas
small_text
dkpro-cassis
spacy
pycaprio